pythonProject7/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
numpy>=1.26.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
scikit-learn>=1.3.0
webcolors>=1.13
scipy>=1.11.0
